requirements.txt
fastapi
uvicorn
pandas
scikit-learn==1.8.0
joblib
pyyaml